+++ src/FunFair.CodeAnalysis/ProhibitedMethodWithStrictParametersInvocationDiagnosticsAnalyzer.cs
@@ -21,17 +21,13 @@ namespace FunFair.CodeAnalysis
         private static readonly ProhibitedMethodsSpec[] ForcedMethods =
         {
             new(ruleId: Rules.RuleDontUseSubstituteReceivedWithZeroNumberOfCalls, title: "Avoid use of received with zero call count", message:
-                "Only use Received with expected call count greater than 0, use DidNotReceived instead if 0 call received expected", sourceClass:
-                "NSubstitute.SubstituteExtensions", forcedMethod: "Received", new[]
-                                                                              {
-                                                                                  new[]
-                                                                                  {
-                                                                                      new ParameterSpec(
-                                                                                          name: "requiredNumberOfCalls",
-                                                                                          type: "NumericLiteralExpression",
-                                                                                          value: "0")
-                                                                                  }
-                                                                              })
+                "Only use Received with expected call count greater than 0, use DidNotReceived instead if 0 call received expected", sourceClass: "NSubstitute.SubstituteExtensions", forcedMethod:
+                "Received", new[] {new[] {new ParameterSpec(name: "requiredNumberOfCalls", type: "NumericLiteralExpression", value: "0")}}),
+            new(ruleId: Rules.RuleDontUseConfigurationBuilderAddJsonFileWithReload, title: "Avoid use of reloadOnChange with value true", message:
+                "Only use AddJsonFile with reloadOnChange set to false", sourceClass: "Microsoft.Extensions.Configuration.JsonConfigurationExtensions", forcedMethod: "AddJsonFile", new[]
+                    {
+                        new[] {new ParameterSpec(name: "reloadOnChange", type: "TrueLiteralExpression", value: "true")}
+                    })
         };
 
         /// <inheritdoc />
@@ -56,18 +52,14 @@ namespace FunFair.CodeAnalysis
         {
             void LookForForcedMethods(SyntaxNodeAnalysisContext syntaxNodeAnalysisContext)
             {
-                InvocationExpressionSyntax[] invocations = syntaxNodeAnalysisContext.Node.DescendantNodesAndSelf()
-                                                                                    .OfType<InvocationExpressionSyntax>()
-                                                                                    .ToArray();
-
-                foreach (InvocationExpressionSyntax invocation in invocations)
+                if (syntaxNodeAnalysisContext.Node is InvocationExpressionSyntax invocation)
                 {
                     IMethodSymbol? memberSymbol = MethodSymbolHelper.FindInvokedMemberSymbol(invocation: invocation, syntaxNodeAnalysisContext: syntaxNodeAnalysisContext);
 
                     // check if there is at least one rule that correspond to invocation method
                     if (memberSymbol == null)
                     {
-                        continue;
+                        return;
                     }
 
                     Mapping mapping = new(methodName: memberSymbol.Name, SymbolDisplay.ToDisplayString(memberSymbol.ContainingType));
@@ -84,7 +76,7 @@ namespace FunFair.CodeAnalysis
                 }
             }
 
-            compilationStartContext.RegisterSyntaxNodeAction(action: LookForForcedMethods, SyntaxKind.MethodDeclaration);
+            compilationStartContext.RegisterSyntaxNodeAction(action: LookForForcedMethods, SyntaxKind.InvocationExpression);
         }
 
         /// <summary>
@@ -129,12 +121,7 @@ namespace FunFair.CodeAnalysis
 
         private sealed class ProhibitedMethodsSpec
         {
-            public ProhibitedMethodsSpec(string ruleId,
-                                         string title,
-                                         string message,
-                                         string sourceClass,
-                                         string forcedMethod,
-                                         IEnumerable<IEnumerable<ParameterSpec>> bannedSignatures)
+            public ProhibitedMethodsSpec(string ruleId, string title, string message, string sourceClass, string forcedMethod, IEnumerable<IEnumerable<ParameterSpec>> bannedSignatures)
             {
                 this.SourceClass = sourceClass;
                 this.ForcedMethod = forcedMethod;